server_backend/order-worker/requirements.txt
flask
pika
psycopg2-binary
python-json-logger
flask-cors
